server/requirements.txt
langchain-google-genai==0.0.8
langchain-core==0.1.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.4
email-validator==2.1.0
jinja2==3.1.2
pytest
httpx
unstructured
python-magic
